@@ -1,35 +1,34 @@
 package com.expressapps.presentexpress;
 
+import android.animation.Animator;
+import android.animation.AnimatorListenerAdapter;
+import android.animation.ValueAnimator;
 import android.annotation.SuppressLint;
 
-import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.fragment.app.Fragment;
 
-import android.app.Activity;
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.graphics.drawable.GradientDrawable;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
-import android.transition.Slide;
 import android.view.GestureDetector;
+import android.view.Gravity;
 import android.view.MotionEvent;
 import android.view.View;
+import android.view.ViewGroup;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
 import android.widget.Toast;
 
-import java.util.ArrayList;
+import com.expressapps.presentexpress.helper.Funcs;
+import com.expressapps.presentexpress.helper.Slide;
+import com.expressapps.presentexpress.helper.Transition;
+import com.expressapps.presentexpress.helper.TransitionCategory;
+import com.expressapps.presentexpress.helper.TransitionDirection;
+
+import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
 
-/**
- * An example full-screen activity that shows and hides the system UI (i.e.
- * status bar and navigation/system bar) with user interaction.
- */
 public class SlideshowActivity extends AppCompatActivity implements View.OnClickListener {
     private static final int SWIPE_MIN_DISTANCE = 120;
     private static final int SWIPE_MAX_OFF_PATH = 250;
@@ -37,52 +36,49 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
     private GestureDetector gestureDetector;
     View.OnTouchListener gestureListener;
 
-    Timer timer = new Timer();
-    int currentSlide = 0;
-
     /**
-     * Some older devices needs a small delay between UI widget updates
+     * Some older devices need a small delay between UI widget updates
      * and a change of the status and navigation bar.
      */
     private static final int UI_ANIMATION_DELAY = 300;
     private final Handler mHideHandler = new Handler();
-    private ImageView mContentView;
+    private FrameLayout mContentView;
 
     private final Runnable mHidePart2Runnable = new Runnable() {
         @SuppressLint("InlinedApi")
         @Override
         public void run() {
             // Delayed removal of status and navigation bar
-
-            // Note that some of these constants are new as of API 16 (Jelly Bean)
-            // and API 19 (KitKat). It is safe to use them, as they are inlined
-            // at compile-time and do nothing on earlier devices.
             mContentView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LOW_PROFILE
-                    | View.SYSTEM_UI_FLAG_FULLSCREEN
-                    | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                    | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                    | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
+                | View.SYSTEM_UI_FLAG_FULLSCREEN
+                | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
+                | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+                | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
+                | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
         }
     };
 
-    private final Runnable mShowPart2Runnable = new Runnable() {
-        @Override
-        public void run() {
-            // Delayed display of UI elements
-            ActionBar actionBar = getSupportActionBar();
-            if (actionBar != null) {
-                actionBar.show();
-            }
-        }
+    private final Runnable mShowPart2Runnable = () -> {
+        // Delayed display of UI elements
+        ActionBar actionBar = getSupportActionBar();
+        if (actionBar != null) actionBar.show();
     };
+
     private boolean mVisible;
-    private final Runnable mHideRunnable = new Runnable() {
-        @Override
-        public void run() {
-            hide();
-        }
-    };
+    private final Runnable mHideRunnable = this::hide;
+
+    Timer timer = new Timer();
+    int currentSlide = -1;
+    private boolean transitionRunning = false;
+    private boolean loopOn;
+    private boolean useTimings;
+    private List<Slide> slides;
+
+    private ImageView photoImg;
+    private ImageView photoImgOther;
+    private FrameLayout photoGrid;
+    private FrameLayout photoGridOther;
+    private ValueAnimator valAnimator;
 
     @SuppressLint("ClickableViewAccessibility")
     @Override
@@ -90,23 +86,327 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_fullscreen);
 
-        FrameLayout background = findViewById(R.id.back);
-        background.setBackgroundColor(MainActivity.backColor);
-
         mVisible = true;
-        mContentView = findViewById(R.id.fullscreen_content);
+        mContentView = findViewById(R.id.back);
+
+        photoGrid = findViewById(R.id.photo_grid_1);
+        photoGridOther = findViewById(R.id.photo_grid_2);
+        photoImg = findViewById(R.id.photo_img_1);
+        photoImgOther = findViewById(R.id.photo_img_2);
+
+        photoGrid.setBackgroundColor(MainActivity.slideshow.info.getBackColour());
+        photoGridOther.setBackgroundColor(MainActivity.slideshow.info.getBackColour());
+
+        loopOn = MainActivity.slideshow.info.loop;
+        useTimings = MainActivity.slideshow.info.useTimings;
+        slides = MainActivity.slideshow.slides;
 
         gestureDetector = new GestureDetector(this, new MyGestureDetector());
-        gestureListener = new View.OnTouchListener() {
-            public boolean onTouch(View v, MotionEvent event) {
-                return gestureDetector.onTouchEvent(event);
-            }
-        };
+        gestureListener = (v, event) -> gestureDetector.onTouchEvent(event);
 
         mContentView.setOnClickListener(SlideshowActivity.this);
         mContentView.setOnTouchListener(gestureListener);
+    }
 
-        mContentView.setImageBitmap((Bitmap) MainActivity.AllSlides.get(0).get("bmp"));
+    @Override
+    protected void onPostCreate(Bundle savedInstanceState) {
+        super.onPostCreate(savedInstanceState);
+
+        // Trigger the initial hide() shortly after the activity has been
+        // created, to briefly hint to the user that UI controls
+        // are available.
+        delayedHide();
+        mHideHandler.postDelayed(this::loadNext, 200);
+    }
+
+    private void loadNext() {
+        stopTimer();
+        currentSlide++;
+
+        if (transitionRunning) {
+            stopTransition();
+
+            if (currentSlide >= slides.size() && !loopOn) {
+                endSlideshow();
+            } else {
+                photoImg.setImageBitmap(slides.get(currentSlide).bitmap);
+                photoGrid.setAlpha(1);
+                transitionFinished();
+            }
+            return;
+        }
+
+        if (currentSlide >= slides.size()) {
+            if (loopOn) {
+                loadStart();
+            } else {
+                endSlideshow();
+            }
+        } else {
+            ImageView tempImg = photoImg;
+            photoImg = photoImgOther;
+            photoImgOther = tempImg;
+
+            FrameLayout tempGrid = photoGrid;
+            photoGrid = photoGridOther;
+            photoGridOther = tempGrid;
+
+            photoImg.setImageBitmap(slides.get(currentSlide).bitmap);
+            photoGrid.setAlpha(1);
+            loadTransition();
+        }
+    }
+
+    private void loadPrevious() {
+        if (currentSlide > 0) {
+            currentSlide -= 2;
+            loadNext();
+
+        } else {
+            findViewById(R.id.photo_grid_1).setAlpha(0);
+            findViewById(R.id.photo_grid_2).setAlpha(0);
+            ((ImageView)findViewById(R.id.photo_img_1)).setImageBitmap(null);
+            ((ImageView)findViewById(R.id.photo_img_2)).setImageBitmap(null);
+
+            transitionRunning = false;
+            loadStart();
+        }
+    }
+
+    private void loadStart() {
+        currentSlide = -1;
+        loadNext();
+    }
+
+    @SuppressLint({"CutPasteId", "RtlHardcoded"})
+    private void loadTransition() {
+        Transition trans = slides.get(currentSlide).transition;
+        stopTransition();
+
+        if (slides.get(currentSlide).transition.getCategory() == TransitionCategory.UNCOVER) {
+            photoGrid.setTranslationZ(1);
+            photoGridOther.setTranslationZ(2);
+        }
+
+        switch (trans.getType()) {
+            case FADE:
+                photoGrid.setAlpha(0f);
+                photoGrid.animate().alpha(1).setDuration(trans.getDurationMs())
+                    .withEndAction(this::transitionFinished);
+                break;
+
+            case FADE_THROUGH_BLACK:
+                photoGridOther.setAlpha(1f);
+                photoGridOther.animate().alpha(0).setDuration(trans.getDurationMs() / 2);
+
+                photoGrid.setAlpha(0f);
+                photoGrid.animate().alpha(1).setStartDelay(trans.getDurationMs() / 2)
+                    .setDuration(trans.getDurationMs() / 2).withEndAction(this::transitionFinished);
+                break;
+
+            case PUSH_LEFT:
+            case PUSH_RIGHT:
+                photoGridOther.setTranslationX(0);
+                photoGridOther.animate()
+                    .translationX(trans.getDirection() == TransitionDirection.LEFT ?
+                        photoGrid.getWidth() : -photoGrid.getWidth())
+                    .setDuration(trans.getDurationMs());
+
+                photoGrid.setTranslationX(trans.getDirection() == TransitionDirection.LEFT ?
+                    -photoGrid.getWidth() : photoGrid.getWidth());
+                photoGrid.animate().translationX(0).setDuration(trans.getDurationMs())
+                    .withEndAction(this::transitionFinished);
+                break;
+
+            case PUSH_TOP:
+            case PUSH_BOTTOM:
+                photoGridOther.setTranslationY(0);
+                photoGridOther.animate()
+                    .translationY(trans.getDirection() == TransitionDirection.TOP ?
+                        photoGrid.getHeight() : -photoGrid.getHeight())
+                    .setDuration(trans.getDurationMs());
+
+                photoGrid.setTranslationY(trans.getDirection() == TransitionDirection.TOP ?
+                    -photoGrid.getHeight() : photoGrid.getHeight());
+                photoGrid.animate().translationY(0).setDuration(trans.getDurationMs())
+                    .withEndAction(this::transitionFinished);
+                break;
+
+            case WIPE_LEFT:
+                photoImg.getLayoutParams().width = findViewById(R.id.back).getWidth();
+
+                valAnimator = ValueAnimator.ofInt(0, findViewById(R.id.back).getWidth());
+                valAnimator.setDuration(trans.getDurationMs());
+                valAnimator.addUpdateListener(animation -> {
+                    photoGrid.getLayoutParams().width = (int) animation.getAnimatedValue();
+                    photoGrid.requestLayout();
+                });
+                valAnimator.addListener(new AnimatorListenerAdapter() {
+                    @Override
+                    public void onAnimationEnd(Animator animation) {
+                        transitionFinished();
+                    }
+                });
+                valAnimator.start();
+                break;
+
+            case WIPE_RIGHT:
+                photoGrid.setLayoutParams(new FrameLayout.LayoutParams(
+                    0, ViewGroup.LayoutParams.MATCH_PARENT, Gravity.RIGHT));
+                photoImg.setLayoutParams(new FrameLayout.LayoutParams(
+                    findViewById(R.id.back).getWidth(), ViewGroup.LayoutParams.MATCH_PARENT, Gravity.LEFT));
+
+                valAnimator = ValueAnimator.ofInt(0, findViewById(R.id.back).getWidth());
+                valAnimator.setDuration(trans.getDurationMs());
+                valAnimator.addUpdateListener(animation -> {
+                    photoGrid.getLayoutParams().width = (int) animation.getAnimatedValue();
+                    photoImg.setTranslationX(-findViewById(R.id.back).getWidth() + (int) animation.getAnimatedValue());
+                    photoGrid.requestLayout();
+                });
+                valAnimator.addListener(new AnimatorListenerAdapter() {
+                    @Override
+                    public void onAnimationEnd(Animator animation) {
+                        transitionFinished();
+                    }
+                });
+                valAnimator.start();
+                break;
+
+            case WIPE_TOP:
+                photoImg.getLayoutParams().height = findViewById(R.id.back).getHeight();
+
+                valAnimator = ValueAnimator.ofInt(0, findViewById(R.id.back).getHeight());
+                valAnimator.setDuration(trans.getDurationMs());
+                valAnimator.addUpdateListener(animation -> {
+                    photoGrid.getLayoutParams().height = (int) animation.getAnimatedValue();
+                    photoGrid.requestLayout();
+                });
+                valAnimator.addListener(new AnimatorListenerAdapter() {
+                    @Override
+                    public void onAnimationEnd(Animator animation) {
+                        transitionFinished();
+                    }
+                });
+                valAnimator.start();
+                break;
+
+            case WIPE_BOTTOM:
+                photoGrid.setLayoutParams(new FrameLayout.LayoutParams(
+                    ViewGroup.LayoutParams.MATCH_PARENT, 0, Gravity.BOTTOM));
+                photoImg.setLayoutParams(new FrameLayout.LayoutParams(
+                    ViewGroup.LayoutParams.MATCH_PARENT, findViewById(R.id.back).getHeight(), Gravity.TOP));
+
+                valAnimator = ValueAnimator.ofInt(0, findViewById(R.id.back).getHeight());
+                valAnimator.setDuration(trans.getDurationMs());
+                valAnimator.addUpdateListener(animation -> {
+                    photoGrid.getLayoutParams().height = (int) animation.getAnimatedValue();
+                    photoImg.setTranslationY(-findViewById(R.id.back).getHeight() + (int) animation.getAnimatedValue());
+                    photoGrid.requestLayout();
+                });
+                valAnimator.addListener(new AnimatorListenerAdapter() {
+                    @Override
+                    public void onAnimationEnd(Animator animation) {
+                        transitionFinished();
+                    }
+                });
+                valAnimator.start();
+                break;
+
+            case UNCOVER_LEFT:
+            case UNCOVER_RIGHT:
+                photoGridOther.setTranslationX(0);
+                photoGridOther.animate()
+                    .translationX(trans.getDirection() == TransitionDirection.LEFT ?
+                        findViewById(R.id.back).getWidth() : -findViewById(R.id.back).getWidth())
+                    .setDuration(trans.getDurationMs()).withEndAction(this::transitionFinished);
+                break;
+
+            case UNCOVER_TOP:
+            case UNCOVER_BOTTOM:
+                photoGridOther.setTranslationY(0);
+                photoGridOther.animate()
+                    .translationY(trans.getDirection() == TransitionDirection.TOP ?
+                        findViewById(R.id.back).getHeight() : -findViewById(R.id.back).getHeight())
+                    .setDuration(trans.getDurationMs()).withEndAction(this::transitionFinished);
+                break;
+
+            case COVER_LEFT:
+            case COVER_RIGHT:
+                photoGrid.setTranslationX(trans.getDirection() == TransitionDirection.LEFT ?
+                    -findViewById(R.id.back).getWidth() : findViewById(R.id.back).getWidth());
+                photoGrid.animate().translationX(0).setDuration(trans.getDurationMs())
+                    .withEndAction(this::transitionFinished);
+                break;
+
+            case COVER_TOP:
+            case COVER_BOTTOM:
+                photoGrid.setTranslationY(trans.getDirection() == TransitionDirection.TOP ?
+                    -findViewById(R.id.back).getHeight() : findViewById(R.id.back).getHeight());
+                photoGrid.animate().translationY(0).setDuration(trans.getDurationMs())
+                    .withEndAction(this::transitionFinished);
+                break;
+
+            case NONE:
+                transitionFinished();
+                return;
+        }
+        transitionRunning = true;
+    }
+
+    private void transitionFinished() {
+        transitionRunning = false;
+        stopTransition();
+
+        if (useTimings) {
+            timer.schedule(new TimerTask() {
+                @Override
+                public void run() {
+                    SlideshowActivity.this.runOnUiThread(() -> loadNext());
+                }
+            }, Math.round(slides.get(currentSlide).getTiming() * 1000));
+        }
+    }
+
+    private void stopTransition() {
+        photoGrid.animate().cancel();
+        photoGridOther.animate().cancel();
+        if (valAnimator != null) {
+            valAnimator.removeAllListeners();
+            valAnimator.cancel();
+        }
+
+        // restore defaults
+        photoGrid.setTranslationZ(2);
+        photoGridOther.setTranslationZ(1);
+
+        photoGrid.setTranslationX(0);
+        photoGridOther.setTranslationX(0);
+        photoGrid.setTranslationY(0);
+        photoGridOther.setTranslationY(0);
+
+        photoImg.setTranslationX(0);
+        photoImgOther.setTranslationX(0);
+        photoImg.setTranslationY(0);
+        photoImgOther.setTranslationY(0);
+
+        photoImg.setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT, Gravity.CENTER));
+        photoImgOther.setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT, Gravity.CENTER));
+
+        photoGrid.setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
+        photoGridOther.setLayoutParams(new FrameLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT));
+    }
+
+    private void endSlideshow() {
+        Funcs.newMessage(getApplicationContext(), R.string.end, Toast.LENGTH_SHORT);
+        timer.cancel();
+        timer.purge();
+        finish();
+    }
+
+    private void stopTimer() {
+        timer.cancel();
+        timer.purge();
+        timer = new Timer();
     }
 
     @Override
@@ -114,6 +414,7 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
 
     @Override
     public void onBackPressed() {
+        stopTransition();
         timer.cancel();
         timer.purge();
         super.onBackPressed();
@@ -125,41 +426,11 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
             try {
                 if (Math.abs(e1.getY() - e2.getY()) > SWIPE_MAX_OFF_PATH) return false;
 
-                if(e1.getX() - e2.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY) {
-                    timer.cancel();
-                    timer.purge();
-                    timer = new Timer();
-
-                    currentSlide++;
-                    if (currentSlide >= MainActivity.AllSlides.size()) {
-                        if (MainActivity.loop) {
-                            currentSlide = 0;
-                            mContentView.setImageBitmap((Bitmap) MainActivity.AllSlides.get(0).get("bmp"));
-                            if (MainActivity.timings) next();
-
-                        } else {
-                            Toast toast = Toast.makeText(getApplicationContext(), R.string.end, Toast.LENGTH_SHORT);
-                            toast.show();
-                            timer.cancel();
-                            timer.purge();
-                            finish();
-                        }
-
-                    } else {
-                        mContentView.setImageBitmap((Bitmap) MainActivity.AllSlides.get(currentSlide).get("bmp"));
-                        if (MainActivity.timings) next();
-                    }
-
+                if (e1.getX() - e2.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY) {
+                    loadNext();
                 } else if (e2.getX() - e1.getX() > SWIPE_MIN_DISTANCE && Math.abs(velocityX) > SWIPE_THRESHOLD_VELOCITY) {
-                    timer.cancel();
-                    timer.purge();
-                    timer = new Timer();
-
-                    if (currentSlide > 0) currentSlide--;
-                    mContentView.setImageBitmap((Bitmap) MainActivity.AllSlides.get(currentSlide).get("bmp"));
-                    if (MainActivity.timings) next();
+                    loadPrevious();
                 }
-
             } catch (Exception ignored) {}
             return false;
         }
@@ -174,50 +445,6 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
         public boolean onDown(MotionEvent e) {
             return true;
         }
-    }
-
-    @Override
-    protected void onPostCreate(Bundle savedInstanceState) {
-        super.onPostCreate(savedInstanceState);
-
-        // Trigger the initial hide() shortly after the activity has been
-        // created, to briefly hint to the user that UI controls
-        // are available.
-        delayedHide(100);
-        if (MainActivity.timings) next();
-    }
-
-    private void next() {
-        timer.schedule(new TimerTask() {
-            @Override
-            public void run() {
-                SlideshowActivity.this.runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        currentSlide++;
-                        if (currentSlide >= MainActivity.AllSlides.size()) {
-                            if (MainActivity.loop) {
-                                currentSlide = 0;
-                                mContentView.setImageBitmap((Bitmap) MainActivity.AllSlides.get(0).get("bmp"));
-                                if (MainActivity.timings) next();
-
-                            } else {
-                                Toast toast = Toast.makeText(getApplicationContext(), R.string.end, Toast.LENGTH_SHORT);
-                                toast.show();
-                                timer.cancel();
-                                timer.purge();
-                                finish();
-                            }
-
-                        } else {
-                            mContentView.setImageBitmap((Bitmap) MainActivity.AllSlides.get(currentSlide).get("bmp"));
-                            next();
-                        }
-                    }
-                });
-
-            }
-        }, Math.round(((Double) MainActivity.AllSlides.get(currentSlide).get("timing")) * 1000));
     }
 
     private void toggle() {
@@ -244,7 +471,7 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
     private void show() {
         // Show the system bar
         mContentView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION);
+             | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION);
         mVisible = true;
 
         // Schedule a runnable to display UI elements after a delay
@@ -256,8 +483,8 @@ public class SlideshowActivity extends AppCompatActivity implements View.OnClick
      * Schedules a call to hide() in delay milliseconds, canceling any
      * previously scheduled calls.
      */
-    private void delayedHide(int delayMillis) {
+    private void delayedHide() {
         mHideHandler.removeCallbacks(mHideRunnable);
-        mHideHandler.postDelayed(mHideRunnable, delayMillis);
+        mHideHandler.postDelayed(mHideRunnable, 100);
     }
 }
