@@ -4,25 +4,22 @@ import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.graphics.drawable.ColorDrawable;
-import android.graphics.drawable.GradientDrawable;
 import android.net.Uri;
 import android.os.Bundle;
 
-import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
+import com.expressapps.presentexpress.helper.Funcs;
 import com.google.firebase.analytics.FirebaseAnalytics;
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
 import android.view.View;
-import android.widget.Button;
 import android.widget.TextView;
 
+import java.util.Objects;
+
 public class AboutActivity extends AppCompatActivity {
-    private Toolbar mToolbar;
-    private FirebaseAnalytics mFirebaseAnalytics;
+    private FirebaseAnalytics mFA;
 
     @SuppressLint("SetTextI18n")
     @Override
@@ -30,23 +27,21 @@ public class AboutActivity extends AppCompatActivity {
         setTheme(R.style.AppTheme_GradientStatusBar);
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_about);
-        mToolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(mToolbar);
-        getSupportActionBar().setTitle(R.string.about_present);
-        mFirebaseAnalytics = FirebaseAnalytics.getInstance(this);
+        setSupportActionBar(findViewById(R.id.toolbar));
+        Objects.requireNonNull(getSupportActionBar()).setTitle(R.string.app_name);
+        mFA = FirebaseAnalytics.getInstance(this);
 
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setDisplayShowHomeEnabled(true);
 
         try {
             PackageInfo pInfo = getApplicationContext().getPackageManager()
-                    .getPackageInfo(getApplicationContext().getPackageName(), 0);
+                .getPackageInfo(getApplicationContext().getPackageName(), 0);
             ((TextView)findViewById(R.id.version)).setText("v" + pInfo.versionName);
 
         } catch (PackageManager.NameNotFoundException e) {
-            ((TextView)findViewById(R.id.version)).setText(R.string.unknown_version);
+            ((TextView) findViewById(R.id.version)).setText(R.string.unknown_version);
         }
-
     }
 
     @Override
@@ -57,24 +52,15 @@ public class AboutActivity extends AppCompatActivity {
 
     public void onWebsiteClick(View v) {
         Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://www.johnjds.co.uk/express"));
-        newEventLog("webOpen", "Website button clicked");
+        Funcs.newEventLog(mFA, "webOpen", "Website button clicked");
         startActivity(browserIntent);
         finish();
     }
 
     public void onHelpClick(View v) {
         Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse("https://express.johnjds.co.uk/present/help"));
-        newEventLog("helpOpen", "Help button clicked");
+        Funcs.newEventLog(mFA, "helpOpen", "Help button clicked");
         startActivity(browserIntent);
         finish();
-    }
-
-    private void newEventLog(String id, String type) {
-        try {
-            Bundle bundle = new Bundle();
-            bundle.putString(FirebaseAnalytics.Param.ITEM_ID, id);
-            bundle.putString(FirebaseAnalytics.Param.CONTENT_TYPE, type);
-            mFirebaseAnalytics.logEvent(FirebaseAnalytics.Event.SELECT_CONTENT, bundle);
-        } catch (Exception ignored) {}
     }
 }
